backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-ai[openai]
pydantic-ai[anthropic]
python-dotenv
openai
anthropic
